Don't crash when app is actived twice.
Gtk will call the activate signal again if the binary
is executed while the app is already running.  Ideally
one would handle it in a better way, such as by bringing
the main application window forward.  This commit
is a first step in that direction, at least making sure
that the app doesn't crash.